requirements.txt
python-pptx==0.6.23
pandas==2.2.2
numpy==1.26.4
streamlit==1.37.0